[Engine] Fixed obscure pid file bug

+++ src/engine/platform/Mac/systemMac.mm
@@ -1094,7 +1094,14 @@ idSystemLocal::PIDIsRunning
 ==============
 */
 bool idSystemLocal::PIDIsRunning(sint pid) {
-    return kill(pid, 0) == 0;
+	// For kill(), pid <= 0 has special meanings so
+	// we can't really use it to check if something
+	// is actually running.
+    if(pid <= 0) {
+        return qfalse;
+    } else {
+        return kill(pid, 0) == 0;
+    }
 }
 
 /*
